improve build system, switch to clang, intptr_t is a thing whoa

+++ src/plugin-callback.cpp
@@ -14,7 +14,7 @@ extern Emu6502 cpu;
 void plugin_callback_handler(PluginCallbackType type, void *arg) {
   switch (type) {
     case EMU_EXIT:
-      exit_status = (int)arg;
+      exit_status = (int)(intptr_t)arg;
       is_running = false;
       break;
